requirements.txt
tensorflow==2.4.0rc1
numpy~=1.19.4
matplotlib~=3.3.3
pillow~=8.0.1
tqdm~=4.51.0